+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Evaluating RC task card and associated LERC model card
 - Compatibility with PyTorch 1.8
+- Dataset reader for SuperGLUE BoolQ
 
 ### Changed